refactor to control path change reactivity separately in Link and useNavigate
