+++ packages/solid-router/src/link.tsx
@@ -6,16 +6,18 @@ import {
   deepEqual,
   exactPathTest,
   functionalUpdate,
+  last,
   preloadWarning,
   removeTrailingSlash,
 } from '@tanstack/router-core'
 import { Dynamic } from 'solid-js/web'
+import { createMemo } from 'solid-js'
+import { useMatch } from './useMatch'
 import { useRouterState } from './useRouterState'
 import { useRouter } from './useRouter'
 
 import { useIntersectionObserver } from './utils'
 
-import { useActiveLocation } from './useActiveLocation'
 import type {
   AnyRouter,
   Constrain,
@@ -133,9 +135,29 @@ export function useLinkProps<
     select: (s) => s.location.searchStr,
   })
 
-  const { getFromPath, activeLocation } = useActiveLocation()
+  const routerLocation = useRouterState({
+    select: (s) => s.location,
+  })
 
-  const from = getFromPath(options.from)
+  const matchIndex = useMatch({
+    strict: false,
+    select: (match) => match.index,
+  })
+
+  const from = createMemo(() => {
+    const activeLocationMatches = router.matchRoutes(routerLocation(), {
+      _buildLocation: false,
+      },
+    )
+
+    const activeLocationMatch = last(activeLocationMatches)
+
+    return (
+      options.from ??
+      activeLocationMatch?.fullPath ??
+      router.state.matches[matchIndex()]!.fullPath
+    )
+  })
 
   const _options = () => {
     return {
@@ -146,7 +168,6 @@ export function useLinkProps<
 
   const next = Solid.createMemo(() => {
     currentSearch()
-    activeLocation()
     return router.buildLocation(_options() as any)
   })
