+++ packages/solid-router/src/useNavigate.tsx
@@ -1,6 +1,7 @@
 import * as Solid from 'solid-js'
+import { last } from '@tanstack/router-core'
 import { useRouter } from './useRouter'
-import { useActiveLocation } from './useActiveLocation'
+import { useMatch } from './useMatch'
 import type {
   AnyRouter,
   FromPathOption,
@@ -17,17 +18,26 @@ export function useNavigate<
 }): UseNavigateResult<TDefaultFrom> {
   const router = useRouter()
 
-  const { getFromPath, setActiveLocation } = useActiveLocation(
-    router.latestLocation,
-  )
+  const matchIndex = useMatch({
+    strict: false,
+    select: (match) => match.index,
+  })
 
   return ((options: NavigateOptions) => {
-    setActiveLocation(router.latestLocation)
-    const from = getFromPath(options.from ?? _defaultOpts?.from)
+    const activeLocationMatches = router.matchRoutes(router.latestLocation, {
+      _buildLocation: false,
+    })
+
+    const activeLocationMatch = last(activeLocationMatches)
+
+    const from = options.from ??
+      _defaultOpts?.from ??
+      activeLocationMatch?.fullPath ??
+      router.state.matches[matchIndex()]!.fullPath
 
     return router.navigate({
       ...options,
-      from: from(),
+      from,
     })
   }) as UseNavigateResult<TDefaultFrom>
 }
